refactor: 금전과 관련된 변수들을 Double에서 BigDecimal로 변경
금전적인 계산의 정밀도를 높이기 위해 금전과 관련된 변수들의 타입을 Double에서 BigDecimal로 변경하였습니다.

+++ src/main/java/com/flab/CommerceCore/payment/domain/entity/Payment.java
@@ -4,6 +4,7 @@ import com.flab.CommerceCore.common.enums.Status;
 import com.flab.CommerceCore.order.domain.entity.Order;
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -16,7 +17,7 @@ public class Payment {
     @JoinColumn(name="order_id")
     private Order order;
 
-    private Double amount;
+    private BigDecimal amount;
 
     @Enumerated(EnumType.STRING)
     private Status status;
